add movie detail pages

+++ movie/mysql_schema.sql
@@ -88,23 +88,12 @@ create table shopping_cart
   primary key (customerID, movieID, storeID)
 );
 
--- create table transactions
--- (
---   purchaseID   INTEGER primary key autoincrement,
---   amount       INTEGER(10) not null,
---   purchaseDate datetime not null,
---   paypalID     VARCHAR(20) not null,
---   customerID   INTEGER references customer,
---   movieID      INTEGER references movie,
---   storeID      INTEGER references store
--- );
-
 create table transaction_detail
 (
   paypalID     VARCHAR(20) references transaction_info,
   movieID      INTEGER references movie,
   amount       INTEGER(10) not null,
-  unitPrice        INTEGER(10) not null,
+  unitPrice    FLOAT not null,
   primary key (paypalID, movieID)
 );
 
@@ -114,7 +103,7 @@ create table transaction_info
   purchaseDate    datetime not null,
   customerID      INTEGER references customer,
   storeID         INTEGER references store,
-  totalPrice      INTEGER(10) not null,
+  totalPrice      FLOAT not null,
   shippingAddress VARCHAR(200) not null,
   -- 0: preparing, 1: shipping, 2: delivered
   status          INTEGER(1) not null,
